NMS-12645: Move test to UserIT
