--- smoke-test/src/test/java/org/opennms/smoketest/AdminUserIT.java
@@ -1,58 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2020 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2020 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.smoketest;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.openqa.selenium.By;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class AdminUserIT extends OpenNMSSeleniumIT {
-
-    private static final Logger LOG = LoggerFactory.getLogger(AdminUserIT.class);
-
-    @Before
-    public void setUp() throws Exception {
-        driver.get(getBaseUrlInternal() + "opennms/account/selfService/newPasswordEntry");
-    }
-
-    @Test
-    public void testChangeAdminPassword() throws Exception {
-
-        enterText(By.cssSelector("input[type=password][name=oldpass]"), "admin");
-        enterText(By.cssSelector("input[type=password][name=pass1]"), "OpenNMS");
-        enterText(By.cssSelector("input[type=password][name=pass2]"), "OpenNMS");
-        driver.findElement(By.cssSelector("button[type=submit]")).click();
-        wait.until(ExpectedConditions.urlContains("newPasswordAction"));
-
-    }
-
-}
