+++ smoke-test/src/test/java/org/opennms/smoketest/UserIT.java
@@ -29,6 +29,7 @@
 package org.opennms.smoketest;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 import java.util.List;
 
@@ -119,6 +120,42 @@ public class UserIT extends OpenNMSSeleniumIT {
         findElementById("users(" + USER_NAME + ").doDelete").click();
         handleAlert("Are you sure you want to delete the user " + USER_NAME + "?");
         assertElementDoesNotExist(By.id(USER_NAME));
+    }
+
+    @Test
+    public void testChangeAdminPasswordWithDifferentPasswords() throws Exception {
+
+        driver.get(getBaseUrlInternal() + "opennms/account/selfService/newPasswordEntry");
+        enterText(By.cssSelector("input[type=password][name=oldpass]"), "admin");
+        enterText(By.cssSelector("input[type=password][name=pass1]"), "OpenNMS");
+        enterText(By.cssSelector("input[type=password][name=pass2]"), "OpenNM");
+        driver.findElement(By.cssSelector("button[type=submit]")).click();
+        try {
+            final Alert alert = wait.until(ExpectedConditions.alertIsPresent());
+            alert.dismiss();
+        } catch (final Exception e) {
+            LOG.debug("Got an exception waiting for a 'wrong password' alert.", e);
+            throw e;
+        }
+    }
+
+    @Test
+    public void testChangeAdminPassword() throws Exception {
+
+        driver.get(getBaseUrlInternal() + "opennms/account/selfService/newPasswordEntry");
+        enterText(By.cssSelector("input[type=password][name=oldpass]"), "admin");
+        enterText(By.cssSelector("input[type=password][name=pass1]"), "OpenNMS");
+        enterText(By.cssSelector("input[type=password][name=pass2]"), "OpenNMS");
+        driver.findElement(By.cssSelector("button[type=submit]")).click();
+        assertTrue(wait.until(pageContainsText("Password successfully changed")));
+
+        // Change password back to admin
+        driver.get(getBaseUrlInternal() + "opennms/account/selfService/newPasswordEntry");
+        enterText(By.cssSelector("input[type=password][name=oldpass]"), "OpenNMS");
+        enterText(By.cssSelector("input[type=password][name=pass1]"), "admin");
+        enterText(By.cssSelector("input[type=password][name=pass2]"), "admin");
+        driver.findElement(By.cssSelector("button[type=submit]")).click();
+        assertTrue(wait.until(pageContainsText("Password successfully changed")));
     }
 
 }
